+++ evaluation/class_metrics.py
@@ -1,7 +1,8 @@
 import csv
 import logging
 import sys
-import time 
+import time
+import pandas as pd
 
 from docopt import docopt
 import numpy as np 
@@ -12,18 +13,20 @@ def main():
     """
 
     # Get the argument 
-    args = docopt("""Compute precission, recall, f_1 and f_0.5..
+    args = docopt("""Create classification report
 
     Usage:
-        class_metrics.py <path_truth> <path_file> 
+        class_metrics.py <path_truth> <path_file>  <path_output> 
 
         <path_truth>    = path to binary gold data (tab-separated)
         <path_file>     = path to file containing words and binary values (tab-separated)
+        <path_output>   = csv file to save performance results
 
     """)
-    
+
     path_truth = args['<path_truth>']
     path_file = args['<path_file>']
+    path_output = args['<path_output>']
 
     logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)
     logging.info(__file__.upper())
@@ -43,19 +46,9 @@ def main():
         for row in reader:
             predictions.append(int(row[1]))
 
-    # Compute metrics
-    precision = metrics.precision_score(truth, predictions, zero_division=0)
-    recall = metrics.recall_score(truth, predictions, zero_division=0)
-    f1 = metrics.f1_score(truth, predictions, zero_division=0)
-    f05 = metrics.fbeta_score(truth, predictions, beta=0.5, zero_division=0)
-
-    precision = round(precision, 3)
-    recall = round(recall, 3)
-    f1 = round(f1, 3)
-    f05 = round(f05, 3)
-
-    # Print output
-    print('\t'.join((str(precision), str(recall), str(f1), str(f05))))
+    clsf_report = pd.DataFrame(metrics.classification_report(y_true=truth, y_pred=predictions,
+                                                             output_dict=True)).transpose()
+    clsf_report.to_csv(path_output, index=True)
 
     logging.info("--- %s seconds ---" % (time.time() - start_time))    
     print("")